// app/backend/src/interfaces/car.ts
export interface Car {
  id: number
  marca: string
  modelo: string
  anoModelo: number
  anoFabricacao: number
  renavan: string
  chassi: string
  image: string
  userId: number
}

export interface CarBody {
  marca: string
  modelo: string
  anoModelo: number
  anoFabricacao: number
  renavan: string
  chassi: string
  image: string
  userId: number
}